add studio tests for graph switcher
