+++ jaseci_studio/mocks/data/detailed_object.json
@@ -1,0 +1,51 @@
+{
+  "name": "test3@mail.com",
+  "kind": "Jaseci Master",
+  "jid": "urn:uuid:1607fce5-c105-423e-b053-3373e394a4d9",
+  "j_timestamp": "2023-02-09T11:06:17.135197",
+  "j_type": "master",
+  "j_parent": null,
+  "j_master": "urn:uuid:1607fce5-c105-423e-b053-3373e394a4d9",
+  "j_access": "private",
+  "j_r_acc_ids": [],
+  "j_rw_acc_ids": [],
+  "caller": null,
+  "head_master_id": null,
+  "sub_master_ids": [],
+  "alias_map": {
+    "active:graph": "urn:uuid:5dc1342b-38d7-47a1-a125-a7588a9e4925",
+    "sentinel:main.jir": "urn:uuid:6d50dc20-1658-4e7c-bc21-cd160dbbd22f",
+    "main.jir:architype:root": "urn:uuid:b88f6d1b-d743-4f7f-81ea-22f2c1a882a4",
+    "main.jir:architype:generic": "urn:uuid:3c353811-6f06-4814-bc9f-fbd54abcbbeb",
+    "main.jir:walker:import_discussion_data": "urn:uuid:b0d42d5c-f66c-4632-8dda-3f4aa8977e55",
+    "main.jir:walker:connect_participants": "urn:uuid:47935194-5f8b-4b7b-8571-b9b6d3c1baa4",
+    "main.jir:walker:connect_participant": "urn:uuid:bbb155db-c119-489b-98c4-9360aca287bf",
+    "main.jir:walker:connect_participant_to_pitch": "urn:uuid:886d3b91-2b47-4dce-ba57-9e082a0cff92",
+    "main.jir:architype:participant_link": "urn:uuid:d6fa9cc0-32e9-431a-8591-8e0844148a7a",
+    "main.jir:architype:discussion_link": "urn:uuid:0347feeb-2433-4cda-8b2d-5dce833f89c6",
+    "main.jir:architype:round_link": "urn:uuid:3e1d8cb1-71f6-4446-a894-18a63bebaba3",
+    "main.jir:architype:pitch_link": "urn:uuid:6166674a-7a7f-4d03-b48a-837ee8ac50e7",
+    "main.jir:architype:room_link": "urn:uuid:b2c67957-8a16-45fa-9439-cfdf84c12ab2",
+    "main.jir:architype:table_link": "urn:uuid:da652079-a4e8-4e2b-81f3-35980a8b1e46",
+    "main.jir:architype:discussions_root": "urn:uuid:8c556904-cb61-4287-892f-6e72741a4d96",
+    "main.jir:architype:discussion": "urn:uuid:5fb66236-b193-4763-8caf-128886d53580",
+    "main.jir:architype:round": "urn:uuid:a0f3451b-64f6-4d06-9501-21082614459f",
+    "main.jir:architype:table": "urn:uuid:4abaa80f-4855-49f8-a801-f028270c9bce",
+    "main.jir:architype:participant": "urn:uuid:270e15b8-bd51-41cb-b8a9-b7e169019cf5",
+    "main.jir:architype:pitch": "urn:uuid:ec1c2cb5-614e-443d-9994-60d800a5b719",
+    "main.jir:architype:viewpoint": "urn:uuid:c12e37f0-b490-4bee-94e9-d51cc4c6d62a",
+    "main.jir:architype:vote": "urn:uuid:24c72585-0b48-41cd-a9d0-d99a9a0254e4",
+    "main.jir:architype:chat": "urn:uuid:8584a4e6-19ed-493c-94a7-7a9df4378d84",
+    "main.jir:architype:room": "urn:uuid:827d45a1-8177-4c99-abc5-40546d716daa",
+    "main.jir:architype:knowledge_graph": "urn:uuid:52934a24-29ce-46c8-b892-3283d394627d",
+    "main.jir:walker:init": "urn:uuid:08540421-c674-48ec-8fca-c0018d43bd96",
+    "active:sentinel": "urn:uuid:6d50dc20-1658-4e7c-bc21-cd160dbbd22f"
+  },
+  "perm_default": "private",
+  "active_gph_id": "urn:uuid:5dc1342b-38d7-47a1-a125-a7588a9e4925",
+  "graph_ids": ["urn:uuid:5dc1342b-38d7-47a1-a125-a7588a9e4925"],
+  "spawned_walker_ids": [],
+  "yielded_walkers_ids": [],
+  "active_snt_id": "urn:uuid:6d50dc20-1658-4e7c-bc21-cd160dbbd22f",
+  "sentinel_ids": ["urn:uuid:6d50dc20-1658-4e7c-bc21-cd160dbbd22f"]
+}
